+++ test/run_and_collect_test.dart
@@ -73,7 +73,7 @@ class ThrowingResolver implements Resolver {
 
 void checkIgnoredLinesInFilesCache(
     Map<String, List<List<int>>?> ignoredLinesInFilesCache) {
-  expect(ignoredLinesInFilesCache.length, 3);
+  expect(ignoredLinesInFilesCache.length, 4);
   final keys = ignoredLinesInFilesCache.keys.toList();
   final testAppKey =
       keys.where((element) => element.endsWith('test_app.dart')).single;